% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone.R
\name{fmlm_isochrone}
\alias{fmlm_isochrone}
\title{fmlm_isochrone}
\usage{
fmlm_isochrone(gtfs, from_lat, from_lon, start_time, end_time, max_fm_time,
  fm_speed = 1.39, max_lm_time = max_fm_time, lm_speed = fm_speed,
  route_pattern = NULL, day = NULL, quiet = FALSE)
}
\arguments{
\item{gtfs}{A set of GTFS data returned from \link{extract_gtfs} or, for more
efficient queries, pre-processed with \link{gtfs_timetable}.}

\item{from_lat}{Latitude coordinate of start location}

\item{from_lon}{Longitude coordinate of start location}

\item{start_time}{Desired departure time at start location, either in seconds
after midnight, a vector of two or three integers (hours, minutes) or (hours,
minutes, seconds), an object of class \link{difftime}, \pkg{hms}, or
\pkg{lubridate}.}

\item{end_time}{End time to calculate isochrone}

\item{max_fm_time}{Maximum amount of time, in seconds, allowed to walk to a
station, before starting a public transport trip}

\item{fm_speed}{The walking speed in m/s to be used when calculating first mile times}

\item{max_lm_time}{Maximum amount of time, in seconds, allowed to walk from a
station, after finishing a public transport trip}

\item{lm_speed}{The walking speed in m/s to be used when calculating last mile times}

\item{route_pattern}{Using only those routes matching given pattern, for
example, "^U" for routes starting with "U" (as commonly used for underground
or subway routes. (Parameter not used at all if \code{gtfs} has already been
prepared with \link{gtfs_timetable}.)}

\item{day}{Day of the week on which to calculate route, either as an
unambiguous string (so "tu" and "th" for Tuesday and Thursday), or a number
between 1 = Sunday and 7 = Saturday. If not given, the current day will be
used. (Not used if \code{gtfs} has already been prepared with
\link{gtfs_timetable}.)}

\item{quiet}{Set to \code{TRUE} to suppress screen messages (currently just
regarding timetable construction).}
}
\value{
An \pkg{sf}-formatted polygon representing the reachable area.
}
\description{
Calculate a single isochrone from a given location and time, returning the area
reachable before the specified \code{end_time}. Reachable in this sense includes a
'first mile' leg, followed by a public transportation trip, followed by a
'last mile' leg.
}
