% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-gtfs.R
\name{interpolate_transfers}
\alias{interpolate_transfers}
\title{interpolate_transfers}
\usage{
interpolate_transfers(gtfs, max_transfer_time, walking_speed = 1.39)
}
\arguments{
\item{gtfs}{A set of GTFS data returned from \link{extract_gtfs}}

\item{max_transfer_time}{The maximum walking time in seconds to allow a transfer between stops}

\item{walking_speed}{The walking speed in m/s to be used when calculating transfer times}
}
\value{
An object of class \code{gtfs}
}
\description{
Interpolates transfers based on allowable time and stop proximity.
}
